Fix: transaction_limit should be amount per transaction, not count
BREAKING CHANGE: Change transaction_limit from Integer to Decimal(10,2)

- transaction_limit now represents maximum USD amount per single transaction
- daily_limit represents total USD amount for all transactions per day
- spending_limit remains as monthly/total spending cap

Backend changes:
- Migration: Change transaction_limit column from INTEGER to NUMERIC(10,2)
- Model: Update Card.transaction_limit from Mapped[int] to Mapped[Decimal]
- Schema: Update CardResponse and CardUpdateLimit to use Decimal
- API: Convert transaction_limit to float when calling AccountPE
- Service: Update AccountPE.update_card_limits signature and docs

Mobile changes:
- Model: Change VirtualCard.transactionLimit from int to double
- Update fromJson, toJson, copyWith to handle double
- API Service: Update updateCardLimit signature
- Provider: Update updateCardLimits signature
- UI: Update dialog labels and descriptions for clarity
- UI: Change keyboard type to accept decimals
- UI: Display as "$500/tx" instead of "100 tx/jour"

+++ mobile/lib/providers/cards_provider.dart
@@ -169,7 +169,7 @@ class CardsProvider with ChangeNotifier {
     required String cardId,
     double? spendingLimit,
     double? dailyLimit,
-    int? transactionLimit,
+    double? transactionLimit,
   }) async {
     _error = null;
